Move backend runtime helpers out of main
