+++ crates/runner/src/backend_runtime.rs
@@ -1,0 +1,50 @@
+use anyhow::Result;
+use supersonic_core::backend::{compiled_backends_display, BackendChoice};
+
+use crate::registry::Backend;
+
+pub(crate) fn resolve_backend(choice: BackendChoice, ordinal: usize) -> Result<Backend> {
+    match choice {
+        BackendChoice::Explicit(backend) => {
+            if !gpu_hal::is_backend_compiled(backend) {
+                anyhow::bail!(
+                    "Requested backend {backend} is not compiled into this build. Compiled backends: [{}]",
+                    compiled_backends_display()
+                );
+            }
+            Ok(backend)
+        }
+        BackendChoice::Auto => {
+            if gpu_hal::is_backend_compiled(Backend::Cuda)
+                && gpu_hal::query_device_info(Backend::Cuda, ordinal).is_ok()
+            {
+                return Ok(Backend::Cuda);
+            }
+            if gpu_hal::is_backend_compiled(Backend::Hip)
+                && kernel_ffi::query_gpu_info(ordinal).is_ok()
+            {
+                return Ok(Backend::Hip);
+            }
+            if gpu_hal::is_backend_compiled(Backend::Metal)
+                && gpu_hal::query_device_info(Backend::Metal, ordinal).is_ok()
+            {
+                return Ok(Backend::Metal);
+            }
+            anyhow::bail!(
+                "No usable GPU backend available for device {ordinal}. Compiled backends: [{}]",
+                compiled_backends_display()
+            )
+        }
+    }
+}
+
+pub(crate) fn resolve_oracle_device(spec: &str, backend: Backend, ordinal: usize) -> String {
+    match spec.trim().to_ascii_lowercase().as_str() {
+        "auto" => match backend {
+            Backend::Cuda => format!("cuda:{ordinal}"),
+            Backend::Hip => "cpu".to_string(),
+            Backend::Metal => "cpu".to_string(),
+        },
+        other => other.to_string(),
+    }
+}
